backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
python-dotenv
pillow
pydantic[email]
pydantic-settings
httpx
